server/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9

SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
PyJWT==2.9.0
passlib==1.7.4
minio==7.2.7
python-dotenv==1.0.1
